# Percentile resampling speed-up without Numba