@btrotta-bom thanks for your PR to speed up the percentile resampling plugin (#1548). There is some reticence here (not sure about at the Bureau) to including Numba in the code as we do not wish to rely on it on the HPC. As such I've had a bash at getting a similar speed up without using Numba, which is this draft PR.

The method here is not as fast as the Numba approach if there are say 100 percentiles, but it is as fast (or faster) if the number of percentiles is smaller (around 20 is typical for us).

I've largely tested this with real data, and I have run out of time to write it up here today, but let me know if you are amenable to this approach and I will finish it up tomorrow. There should be some crossover with the speed up of probabilities to percentiles I imagine (#1578).

Testing:
 - [x] Ran tests and they passed OK
 - [x] Added new tests for the new feature(s)