@bayliffe 2021-11-05
I am not sure what the team's overall desire is regarding Numba, and it's too quiet this Friday to reach a consensus 😄 
Personally I am happy to have Numba as an option and use this approach as the fall back, but we will need @tjtg and @benfitzpatrick to weigh in I imagine.

# Testing
I've co-opted your test code to compare all the methods together:

```
from numba.np.ufunc import parallel
import numpy as np
import time
from scipy import interpolate
from numba import njit, prange
from improver.ensemble_copula_coupling.utilities import flatten_and_interpolate, concatenate_array_with_array_endpoints

input_percentile_size = 20
target_percentile_size = 20

x1 = np.sort(np.random.random(input_percentile_size))
# Construct as 3D shape for use with new method
yorig = np.sort(np.random.random_sample((1000, 1000, len(x1))), axis=1) * 2

# Add end points as expected
x1 = concatenate_array_with_array_endpoints(x1, 0, 1)
yorig = concatenate_array_with_array_endpoints(yorig, 0, 2)

# Reshape to original test shape
y1 = yorig.reshape(1000000, len(x1))
x2 = np.linspace(0, 1, target_percentile_size)
np.random.shuffle(x2)

def slow_interp(x1, y1, x2):
    ans = np.zeros((y1.shape[0], len(x2)))
    for i in range(y1.shape[0]):
        ans[i, :] = np.interp(x2, x1, y1[i, :])
    return ans

@njit(parallel=True)
def fast_interp_same_x(x: np.ndarray, xp: np.ndarray, fp: np.ndarray) -> np.ndarray:
    index = np.searchsorted(xp, x)
    result = np.empty((fp.shape[0], len(x)))
    for row in prange(fp.shape[0]):
        for i, ind in enumerate(index):
            if ind == 0:
                result[row, i] = fp[row, 0]
            elif ind == len(xp):
                result[row, i] = fp[row, -1]
            elif xp[ind] - xp[ind - 1] >= 1e-15:
                result[row, i] = fp[row, ind - 1] + (x[i] - xp[ind - 1]) / (
                    xp[ind] - xp[ind - 1]
                ) * (fp[row, ind] - fp[row, ind - 1])
            else:
                result[row, i] = fp[row, ind - 1]
    return result

t = time.time()
slow_ans = slow_interp(x1, y1, x2)
print("old time: {}".format(time.time() - t))
t = time.time()
fast_ans = fast_interp_same_x(x2, x1, y1)
print("new time: {}".format(time.time() - t))
t = time.time()
vec_ans = flatten_and_interpolate(x2, x1, yorig, max_value=2)
vec_ans.data
print("vec time: {}".format(time.time() - t))

vec_reshape = np.moveaxis(vec_ans, 0, -1)

print("same answer: {}".format((np.isclose(slow_ans, fast_ans)).all()))
print("same answer: {}".format((np.isclose(slow_ans.astype(np.float32).flatten(), vec_reshape.flatten())).all()))
```
which on a run using a 1000x1000x20 type size grid (so 20 source percentiles) with 20 target percentiles gives this kind of result:
```
old time: 3.665600061416626
new time: 0.8652346134185791
vec time: 0.3963656425476074
same answer: True
same answer: True
```
Which looks pretty favourable for this approach. However, running some loops in a notebook to get slightly more solid timing it is clear that the Numba approach remains the fastest by a signficant margin:
```
%%timeit
slow_ans = slow_interp(x1, y1, x2)
slow_ans.data
3.67 s ± 106 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%%timeit
fast_ans = fast_interp_same_x(x2, x1, y1)
fast_ans.data
63.8 ms ± 4.46 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%%timeit
vec_ans = flatten_and_interpolate(x2, x1, yorig, max_value=2)
vec_ans.data
393 ms ± 3.53 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
However, both approaches trounce the original.

So if you are happy @btrotta-bom to wait until the technical leads give us a steer we can decide the way forward using one or a combination of these approaches. Thanks for the comments, they were helpful and drew my attention to the very odd reshaping I was doing all over the place.