@btrotta-bom 2021-11-07
> So if you are happy @btrotta-bom to wait until the technical leads give us a steer we can decide the way forward using one or a combination of these approaches. 

Sounds good!