@bayliffe 2021-11-08
numba won this fight....